requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
drf-nested-routers==0.95.0
gunicorn==23.0.0
packaging==25.0
python-decouple==3.8
sqlparse==0.5.3
